% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{addAttributes2}
\alias{addAttributes2}
\title{addAttributes2}
\arguments{
\item{cg}{The collapsed graph from the function \code{\link{collapseGraph}}.}

\item{data}{A tibble with the GO IDs in column \code{GOID} and community
terms in the column \code{communityTerm}.}
}
\description{
Adds vertex and edge attributes to the collapsed graph AFTER the nodes have
been added back from the uncollapsed graph.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{addAttributes2}
