% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{addBackVertexAndEdges}
\alias{addBackVertexAndEdges}
\title{addBackVertexAndEdges}
\arguments{
\item{cg}{The collapsed graph from the function \code{\link{collapseGraph}}.}

\item{assignment}{A data frame of edges from term to community term with
colnames \emph{from} and \emph{to}.}
}
\description{
Takes a GO graph in igraph format, runs the spin-glass community algorithm,
and collapses the graph by community using the name of the node with the
highest authority score to name each vertex in the community.
}
\examples{
#no examples yet

}
\author{
Jason T. Serviss
}
\keyword{addBackVertexAndEdges}
