% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{calculateZscore}
\alias{calculateZscore}
\title{calculateZscore}
\description{
Calculates the z-score, i.e. scales the expression per gene to a range
of 0 to 1 using \code{\link{deResultsRld}}.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{calculateZscore}
