% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{clusterForOrder}
\alias{clusterForOrder}
\title{clusterForOrder}
\description{
Performs a hierarchical clustering per variable and extracts the ordering.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{clusterForOrder}
