% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{collapseGraph}
\alias{collapseGraph}
\title{collapseGraph}
\arguments{
\item{g}{A GO graph in igraph format.}

\item{spins}{The maximum number of spins allowed. Passed
to \link[igraph]{cluster_spinglass}.}
}
\value{
A list with the first element being the collapsed graph and the
   second element being a summary of the community analysis.
}
\description{
Takes a GO graph in igraph format, runs the spin-glass community algorithm,
and collapses the graph by community using the name of the node with the
highest authority score to name each vertex in the community.
}
\examples{
#no examples yet

}
\author{
Jason T. Serviss
}
\keyword{collapseGraph}
