% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deResultsRld.R
\docType{data}
\name{deResultsRld}
\alias{deResultsRld}
\title{Results from the DESeq2 differential expression analysis.}
\format{data.frame
\describe{
  \item{rownames}{Ensembl IDs}
  \item{colnames}{Sample identification}
}}
\usage{
deResultsRld
}
\value{
A data frame containing the regularized log transformed counts after
   differential expression analysis with
   \href{http://bioconductor.org/packages/DESeq2/}{DESeq2}. The data was
   generated using the \link[DESeq2]{rlogTransformation} function.
}
\description{
Produced with data-raw/runHCT.R runHCT().
}
\details{
Regularized log transformed expression values.
}
\examples{
deResultsRld

}
