% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{else2}
\alias{else2}
\title{else the new way.}
\usage{
else2(else_ret)
}
\arguments{
\item{else_ret}{The else return.}
}
\description{
else the new way.
}
\examples{
#\\href{https://www.r-bloggers.com/a-wrapper-around-nested-ifelse/}
   {R-bloggers}

}
\author{
\href{https://github.com/edwinth/thatssorandom}{Edwin Thoen}
}
\keyword{else2}
