% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{extractGenesFromGO}
\alias{extractGenesFromGO}
\title{extractGenesFromGO}
\description{
Extracts Ensembl IDs from the topGOresult and returns them in a long
data frame.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{extractGenesFromGO}
