% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{g_legend}
\alias{g_legend}
\title{Get ggplot legend.}
\usage{
g_legend(a.ggplot)
}
\arguments{
\item{a.gplot}{A ggplot.}
}
\description{
Get ggplot legend.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{g_legend}
