% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\docType{data}
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Gene nnotation from ensembl.}
\format{data frame
\describe{
 \item{ID}{character. Ensembl ID.}
 \item{chr}{character. Chromosome.}
 \item{start}{int. Transcription start site.}
 \item{end}{int. Transcription end site.}
 \item{strand}{factor. Strand.}
 \item{geneName}{character. Gene name.}
 \item{biotype}{character. Gene biotype.}
 \item{description}{character. Gene description.}
}}
\usage{
geneAnnotation
}
\value{
A data frame with Ensembl gene annotation.
}
\description{
Assembled with data-raw/getAnnotation.R getAnnotation().
}
\details{
Gene Annotation.
}
\examples{
geneAnnotation

}
