% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{if2}
\alias{if2}
\title{if the new way.}
\usage{
if2(if_stat, then)
}
\arguments{
\item{if_stat}{The logical expression.}

\item{then}{The then statment.}
}
\description{
if the new way.
}
\examples{
#\\href{https://www.r-bloggers.com/a-wrapper-around-nested-ifelse/}
   {R-bloggers}

}
\author{
\href{https://github.com/edwinth/thatssorandom}{Edwin Thoen}
}
\keyword{if2}
