% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{ifelse2}
\alias{ifelse2}
\title{ifelse the new way.}
\usage{
ifelse2(...)
}
\arguments{
\item{...}{if2 objects followed by one else2 object.}
}
\description{
ifelse the new way.
}
\examples{
#\\href{https://www.r-bloggers.com/a-wrapper-around-nested-ifelse/}
   {R-bloggers}

}
\author{
\href{https://github.com/edwinth/thatssorandom}{Edwin Thoen}
}
\keyword{ifelse2}
