% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{MAplot}
\alias{MAplot}
\title{MAplot}
\description{
Plot an MA plot (log 2 average expression vs. log2 fold change) using the
results from the DESeq2 differential expression analysis.
}
\examples{

MAplot()

}
\author{
Jason T. Serviss
}
\keyword{MAplot}
