% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{PCA}
\alias{PCA}
\title{PCA}
\arguments{
\item{ntop}{The number of genes with the to include ranked by their variance.}

\item{PCmin}{The first principal component to return.}

\item{PCmax}{The last principal component to return.}

\item{plotType}{A character vector. Can be either \code{2D}, \code{plotly} or
\code{scatter3D}.}
}
\value{
Returns a data frame with the PCA results generated by prcomp(). In
addition a 3d plot is plotted with the 3 principal components specified by
PCmin and PCmax.
}
\description{
Performs principal component analysis.
}
\details{
This function performs PCA using the genes specified by the \code{ntop}
argument and supports 3D plotting of the resulting principal components.
The range of \code{PCmin} to \code{PCmax} must include 3 principal
components.
}
\examples{
pca.result <- PCA()
#to save the plot use:
\dontrun{rgl.postscript("PCAplot","pdf")}

}
\author{
Jason T. Serviss
}
\keyword{PCA}
