% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{PCvar}
\alias{PCvar}
\title{PCvar}
\arguments{
\item{x}{Output from the PCA function.}
}
\description{
Plots the percentage of variance in each principal component
}
\details{
This function allows plotting of the percentage of variance in the
principal components for PCAs generated with the PCA function. A red line is
drawn at the level of variance contributed by each principal component if
each principal component contributed an equal amount of variance.
}
\examples{
pca <- PCA()
PCvar(pca)

}
\author{
Jason T. Serviss
}
\keyword{PCA,}
\keyword{PCvar}
