% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{plotCollapsedGraph}
\alias{plotCollapsedGraph}
\title{plotCollapsedGraph}
\arguments{
\item{cg}{The collapsed graph from the function \code{\link{collapseGraph}}.}
}
\description{
Plots the collapsed graph.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{plotCollapsedGraph}
