% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{plotCondRename}
\alias{plotCondRename}
\title{Rename conditions for plotting}
\arguments{
\item{A}{character vector of conditions to be renamed.}

\item{reps}{Logical. Should the replicate label be included?}
}
\description{
Rename conditions for plotting
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{plotCondRename}
