% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFigures.R
\name{plotFigure}
\alias{plotFigure}
\alias{plotFigure}
\title{plotFigure}
\usage{
plotFigure(figure, ...)
}
\arguments{
\item{figure}{Character; Can be "go_analysis" or "RNAseq"}

\item{...}{additional arguments to pass on.}
}
\value{
spCounts object.
}
\description{
Recreates any figure from scratch.
}
\details{
Include description.
}
\examples{

\dontrun{plotFigure("go_analysis")}

}
\author{
Jason T. Serviss
}
