% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes in community terms.}
\arguments{
\item{data}{The plot data.}

\item{term}{The term to plot.}

\item{n}{The number of genes to plot.}
}
\description{
Plot genes in community terms.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{plotGenes}
