% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{rescale}
\alias{rescale}
\title{rescale}
\description{
Rescales the \code{\link{deResultsRld}} expression per gene to a range
of 0 to 1.
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{rescale}
