% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{runTopGO}
\alias{runTopGO}
\title{runTopGO}
\arguments{
\item{save}{Logical. Indicates if the results should be
saved to ./data/topGOresult.rda.}
}
\description{
Runs GO analysis using topGO.
}
\examples{
\dontrun{runTopGO()}

}
\author{
Jason T. Serviss
}
\keyword{runTopGO}
