% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{TSNE}
\alias{TSNE}
\title{TSNE}
\arguments{
\item{ntop}{The number of genes to include in t-SNE ranked by their variance.}
}
\description{
Runs t-SNE and plots the results.
}
\details{
This function runs t-SNE for all samples and the specified genes using
1-Pearson's correlation as input to the t-SNE algorithm and plots the results.
}
\examples{
TSNE()

}
\author{
Jason T. Serviss
}
