% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcanoPlot}
\arguments{
\item{n}{Numeric. The number of gene names to plot based on their log2 fold
change.}
}
\description{
Plot an volcano plot using the results from the DESeq2 differential
expression analysis using only significant (alpha < 0.05) genes.
}
\examples{

volcanoPlot()

}
\author{
Jason T. Serviss
}
\keyword{volcanoPlot}
