% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{downloadGOgraph}
\alias{downloadGOgraph}
\title{downloadGOgraph}
\arguments{
\item{terms}{Character. A character vector of GO IDs.}
}
\value{
A GO graph in igraph format with vertex attributes \emph{name} set to
   the GO IDs and the edge attribute \emph{weight} set to 1 for all edges.
}
\description{
Downloads the GO graph for a set of GO IDs and converts it to igraph format.
}
\examples{
downloadGOgraph("GO:0006915")

}
\author{
Jason T. Serviss
}
\keyword{downloadGOgraph}
