% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOanalysis.R
\name{plotCondRename}
\alias{plotCondRename}
\title{Rename conditions for plotting}
\description{
Rename conditions for plotting
}
\examples{
#no example yet

}
\author{
Jason T. Serviss
}
\keyword{plotCondRename}
